+++ src/dcl-scene-ui-workaround/resources.ts
